+++ src/core/base.jl
@@ -28,7 +28,10 @@ function run_model(file::String, model_type::Type, optimizer, build_method; kwar
 end
 
 ""
-function run_model(data::Dict{String,<:Any}, model_type::Type, optimizer, build_method; ref_extensions=[], solution_processors=[], multinetwork=false, multiconductor=false, kwargs...)
+function run_model(data::Dict{String,<:Any}, model_type::Type, optimizer, build_method;
+        ref_extensions=[], solution_processors=[], relax_integrality=false,
+        multinetwork=false, multiconductor=false, kwargs...)
+
     if multinetwork != _IM.ismultinetwork(data)
         model_requirement = multinetwork ? "multi-network" : "single-network"
         data_type = _IM.ismultinetwork(data) ? "multi-network" : "single-network"
@@ -46,7 +49,7 @@ function run_model(data::Dict{String,<:Any}, model_type::Type, optimizer, build_
     Memento.debug(_LOGGER, "pm model build time: $(time() - start_time)")
 
     start_time = time()
-    result = optimize_model!(pm, optimizer=optimizer, solution_processors=solution_processors)
+    result = optimize_model!(pm, relax_integrality=relax_integrality, optimizer=optimizer, solution_processors=solution_processors)
     Memento.debug(_LOGGER, "pm model solve and solution time: $(time() - start_time)")
 
     return result